update theme settings components and removed context
